Codex org

See merge request tenzorsoft/toir/toir-backend!706

+++ src/main/java/com/toir/dto/pprplanning/calendar/PprEquipmentCalendarFilter.java
@@ -45,7 +45,7 @@ public record PprEquipmentCalendarFilter(
             Boolean onlyWithWork,
             Boolean includeCancelled
     ) {
-        this(Objects.requireNonNull(year, "year is required"),
+        this(Objects.requireNonNull(year, "year is required").intValue(),
                 page == null ? 0 : page,
                 size == null ? 25 : size,
                 search,
